+++ src/fr/epf/computer/service/CompanyService.java
@@ -9,5 +9,5 @@ public interface CompanyService {
     List<Company> getCompanies();
     List<Company> searchByName(String name);
     void create(Company company);
-    Company findById(int id);
+    Company findById(long id);
 }
